backend/requirements.txt
# Dependencias del backend
Django>=5.2,<6
djangorestframework>=3.15,<4
django-cors-headers>=4.3,<5
djangorestframework-simplejwt>=5.3,<6
django-filter>=24.2,<25
